change to hpp and split float and double into two files
